⚡️ Only import the required function
This change prefers an optimized and stripped-down includes-directory at
the cost of package/function pulling speed. Actually, it is not pulling,
as the pulled package/image is cached and will be re-used for subsequent
functions; it is the extraction overhead.

Another nice side-effect is that individual functions of a package can
be overridden locally, as existing functions are not pulled.

Signed-off-by: Daniel Ruthardt <druthardt@linuxfoundation.org>